+++ go.mod
@@ -1,12 +1,12 @@
-module attendanceManagement
+module github.com/nanato-okajima/attendance_management
 
 go 1.18
 
 require (
-	github.com/go-playground/validator/v10 v10.11.0
+	github.com/go-playground/validator v9.31.0+incompatible
 	github.com/kelseyhightower/envconfig v1.4.0
 	gorm.io/driver/mysql v1.3.4
-	gorm.io/gorm v1.23.7
+	gorm.io/gorm v1.23.8
 )
 
 require (
@@ -16,7 +16,5 @@ require (
 	github.com/jinzhu/inflection v1.0.0 // indirect
 	github.com/jinzhu/now v1.1.4 // indirect
 	github.com/leodido/go-urn v1.2.1 // indirect
-	golang.org/x/crypto v0.0.0-20211215153901-e495a2d5b3d3 // indirect
-	golang.org/x/sys v0.0.0-20210806184541-e5e7981a1069 // indirect
-	golang.org/x/text v0.3.7 // indirect
+	gopkg.in/go-playground/assert.v1 v1.2.1 // indirect
 )
